# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  delivery.productRequest:
    properties:
      name:
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
...
  models.Product:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: number
      rest_id:
        type: integer
    type: object
...
  tools.Error:
    properties:
      error:
        type: string
    type: object
  tools.Message:
    properties:
      message:
        type: object
    type: object
host: localhost:5000
info:
  contact:
    name: API Support
  description: This is a SkyDelivery server for Technopark Project.
  license: {}
  title: Swagger SkyDelivery API
  version: "1.0"
paths:
...
  /product/:prod_id:
    delete:
      consumes:
      - application/json
      description: Updating Product Image
      parameters:
      - description: Product ID
        in: path
        name: prod_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.Error'
      summary: Update Product Image Route
      tags:
      - Product
  /product/:prod_id/image:
    put:
      consumes:
      - multipart/form-data
      description: Updating Product Image
      parameters:
      - description: Product ID
        in: path
        name: prod_id
        required: true
        type: integer
      - description: New product image
        in: formData
        name: ProdReq
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.Error'
      summary: Update Product Image Route
      tags:
      - Product
  /product/:prod_id/update:
    put:
      consumes:
      - application/json
      description: Updating Product
      parameters:
      - description: Product ID
        in: path
        name: prod_id
        required: true
        type: integer
      - description: New product data
        in: body
        name: ProdReq
        required: true
        schema:
          $ref: '#/definitions/delivery.productRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
      summary: Update Product Route
      tags:
      - Product
  /product/prod:id:
    get:
      consumes:
      - application/json
      description: Returning Product Model
      parameters:
      - description: Product ID
        in: path
        name: prod_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tools.Error'
      summary: Get Product Route
      tags:
      - Product
...
  /restaurants/rest:id/product:
    get:
      consumes:
      - application/json
      description: Returning Products List of Restaurant
      parameters:
      - description: Count of elements on page
        in: query
        name: count
        required: true
        type: integer
      - description: Number of page
        in: query
        name: page
        required: true
        type: integer
      - description: Id of restaurant
        in: path
        name: rest_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
      summary: Get Products Of Restaurant Route
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Creating Product
      parameters:
      - description: Restaurant ID
        in: path
        name: rest_id
        required: true
        type: integer
      - description: New product name
        in: formData
        name: Name
        required: true
        type: string
      - description: New product price
        in: formData
        name: Price
        required: true
        type: number
      - description: New product image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
      summary: Create Product Route
      tags:
      - Product
...
swagger: "2.0"